\name{forecast-class}
\docType{class}
\alias{forecast-class}

\title{Class \code{"forecast"}}
\description{
Container class for storing forecast model and it's parameters
}
\section{Slots}{
  \describe{
    \item{\code{java}:}{Object of class \code{"jobjRef"} ~~ }
  }
}

\author{Kostin Andrey <andrey.kostin@portfolioeffect.com>}
\examples{
showClass("forecast")
}
\keyword{PortfolioEffectHFT}
%\concept{high frequency, intraday analytics, market data, portfolio, portfolio management,realtime analytics, risk, risk management, toolbox tools, trading, trading strategies}
\keyword{classes}
