\name{optimizer-class}
\docType{class}
\alias{optimizer-class}
\title{Class \code{"optimizer"}}
\description{
Class for storing optimization goals and constraints.
}
\section{Slots}{
  \describe{
    \item{\code{java}:}{Object of class \code{"jobjRef"} ~~ }
    \item{\code{portfolio}:}{Object of class \code{"jobjRef"} ~~ }
  }
}

\author{Kostin Andrey <andrey.kostin@portfolioeffect.com>}
\examples{
showClass("optimizer")
}
\keyword{PortfolioEffectHFT}
%\concept{high frequency, market data, optimization, realtime analytics, risk, risk management, trading, trading strategies}
\keyword{classes}
