\name{util_fillScheme}
\alias{util_fillScheme}
\title{Fill scheme for charts }
\usage{util_fillScheme()
}
\value{Constructed ggplot2 discrete color scheme.
}
\description{Custom fill scheme for charts based on ggplot2 discrete scale.}
\author{Kostin Andrey <andrey.kostin@portfolioeffect.com>}
\examples{
\dontrun{ 

util_fillScheme()
}}
\keyword{PortfolioEffectHFT}
\keyword{util_fillScheme}